+++ tests/testthat/test-provider-claude.R
@@ -57,9 +57,7 @@ test_that("can use pdfs", {
 
 test_that("can set beta headers", {
   chat <- chat_claude_test(beta_headers = c("a", "b"))
-  provider <- chat$.__enclos_env__$private$provider
-
-  req <- chat_request(provider)
+  req <- chat_request(chat$get_provider())
   expect_equal(req$headers$`anthropic-beta`, c("a", "b"))
 })